refactor: move usePollRegular from index.ts
